refactor: unify both cascade and default modes (#33)
refactor(helpers): modularize cascade mode MutationOberver into
"onAnimationEnd"
