--- lib/component/esm/cascade.mjs
@@ -1,1 +1,0 @@
-import"./typingInterval.mjs";import{o as t}from"./writeEffect.mjs";export default async({elements:e},o)=>{e.forEach((t=>t.currentNode.textContent=""));for(const r of e)await t(r,o);o.dispatch("done")};
