--- lib/component/esm/default.mjs
@@ -1,1 +1,0 @@
-import"./typingInterval.mjs";import{o as t}from"./writeEffect.mjs";import{e}from"./getLongestTextElement.mjs";export default async({elements:s},o)=>{for(const e of s)t(e,o);const r=e(s);new MutationObserver((t=>{t.forEach((t=>{const e="attributes"===t.type,s=!t.target.classList.contains("typing");e&&s&&o.dispatch("done")}))})).observe(r,{attributes:!0,childList:!0,subtree:!0})};
