--- lib/component/esm/getLongestTextElement.mjs
@@ -1,1 +1,0 @@
-const t=(t,e)=>e.text.length-t.text.length,e=e=>e.sort(t)[0].currentNode;export{e};
