--- lib/core/modes/cascade.js
@@ -1,1 +1,0 @@
-import{writeEffect as e}from"../helpers/writeEffect.js";export default async({elements:t},o)=>{t.forEach((e=>e.currentNode.textContent=""));for(const r of t)await e(r,o);o.dispatch("done")};
