+++ lib/core/modes/index.js
@@ -1,1 +1,1 @@
-import{getElements as a}from"../helpers/getElements.js";const e=async(e,s)=>{const{default:t}=(s.loop||s.loopRandom)&&await import("./loop.js")||s.cascade&&await import("./cascade.js")||s.scramble&&await import("./scramble.js")||await import("./default.js"),o=a(e);if(s.delay>0){const{sleep:a}=await import("../helpers/index.js");await a(s.delay),e.classList.remove("delay")}t({node:e,elements:o},s)};export{e as typewriter};
+import{getElements as e}from"../helpers/getElements.js";const t=async(t,a)=>{const{default:o}=(a.loop||a.loopRandom)&&await import("./loop.js")||a.scramble&&await import("./scramble.js")||await import("./typewriter.js"),s=e(t);if(a.delay>0){const{sleep:e}=await import("../helpers/index.js");await e(a.delay),t.classList.remove("delay")}o({node:t,elements:s},a)};export{t as typewriter};
