+++ lib/core/modes/typewriter.js
@@ -1,0 +1,1 @@
+import{writeEffect as e}from"../helpers/writeEffect.js";import{onAnimationEnd as t}from"../helpers/onAnimationEnd.js";export default async({elements:o},s)=>{if(s.cascade)(e=>{e.forEach((e=>e.currentNode.textContent=""))})(o);else{const{getLongestTextElement:e}=await import("../helpers/index.js"),a=e(o);t(a,(()=>s.dispatch("done")))}for(const t of o)s.cascade?await e(t,s):e(t,s);s.cascade&&s.dispatch("done")};
