+++ src/helpers/index.ts
@@ -11,3 +11,4 @@ export * from './isInRange'
 export * from './writeEffect'
 export * from './hideCursorOnAnimationEnd'
 export * from './getLongestTextElement'
+export * from './onAnimationEnd'
