+++ src/modes/index.ts
@@ -4,9 +4,8 @@ import type { TypewriterMainFn } from '@svelte-typewriter/types'
 const typewriter: TypewriterMainFn = async (node, options) => {
 	const { default: mode } =
 		((options.loop || options.loopRandom) && (await import('./loop'))) ||
-		(options.cascade && (await import('./cascade'))) ||
 		(options.scramble && (await import('./scramble'))) ||
-		(await import('./default'))
+		(await import('./typewriter'))
 	const elements = getElements(node)
 	if (options.delay > 0) {
 		const { sleep } = await import('../helpers')
